+++ packages/component/jest.config.js
@@ -1,8 +1,5 @@
 module.exports = {
-  setupFilesAfterEnv: [
-    './utils/setupDotEnv.js',
-    './utils/setupGlobalAgent.js',
-    './utils/setupFetchPolyfill.js'
-  ],
-  testEnvironment: 'node'
+  setupFilesAfterEnv: ['./utils/setupDotEnv.js', './utils/setupGlobalAgent.js', './utils/setupFetchPolyfill.js'],
+  testEnvironment: 'node',
+  transformIgnorePatterns: ['<rootDir>/node_modules/(?!(uuid)/)']
 };