+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [7.1.2] - 2022-09-27
+
 ### Changed
 
 - Bumped dependencies, in PR [#196](https://github.com/compulim/web-speech-cognitive-services/pull/196)